move user_usage in a user service
