+++ backend/models/__init__.py
@@ -3,4 +3,3 @@ from .files import File
 from .settings import (BrainRateLimiting, BrainSettings, ResendSettings,
                        get_documents_vector_store, get_embeddings,
                        get_supabase_client, get_supabase_db)
-from .user_usage import UserUsage
